Add spacing before h2
With the default formatting, it is difficult to differentiate between
h2 and h3 headings. In particular, the h2 headings separate
significant portions of my posts (e.g. "Part 1" vs. "Part 2"). I want
some way to easily scan and see the structure of the post.

Adding a significant spacing before the h2 headings seemed to fulfill
this desire nicely. I don't want to add padding to the first h2
heading, though, so I use a `first-child` selector to reset the
padding to its initial value for the first h2 heading.