Remove categories from posts
I had no idea that categories were added to the url as paths before
the blog date. For now I only plan on doing Advent of Code. However,
if I re-use this for something else, it could be nice to have other
sections. I'll keep the `adventofcode` category but remove everything
else. Using `python` and `coding` would be good as tags but not as
categories I suppose.